// packages/frontend/src/integrations/supabase/types.ts
export type Database = {
  // Allows to automatically instantiate createClient with right options
  // instead of createClient<Database, { PostgrestVersion: 'XX' }>(URL, KEY)
  __InternalSupabase: {
    PostgrestVersion: "13.0.4"
  }
  public: {
    Tables: {
// ...
      preview_builds: {
        Row: {
          created_at: string | null
          id: string
          commit_sha: string | null
          run_id: string | null
          build_number: string | null
          is_seen: boolean
          pr_number: number
          pr_url: string
          preview_url: string
          status: 'pending_review' | 'committed'
        }
        Insert: {
          created_at?: string | null
          id?: string
          commit_sha?: string | null
          run_id?: string | null
          build_number?: string | null
          is_seen?: boolean
          pr_number: number
          pr_url: string
          preview_url: string
          status?: 'pending_review' | 'committed'
        }
        Update: {
          created_at?: string | null
          id?: string
          commit_sha?: string | null
          run_id?: string | null
          build_number?: string | null
          is_seen?: boolean
          pr_number?: number
          pr_url?: string
          preview_url?: string
          status?: 'pending_review' | 'committed'
        }
        Relationships: []
      }
// ...
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      [_ in never]: never
    }
    Enums: {
      [_ in never]: never
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
}